such a large refactor omg - removed loading of .sm and .osu - began groundwork for conversion of .sm and .osu to .yav - added loading and caching for .yav - restructured chart loading into individual difficulties rather than one song and multiple difficulties - organised lots of code into folders better - changed data types for some technical map storage to ~halve ram usage of note data - rewrote sprite loading system to support live skin swapping a bit better - rewrote spritebatch to be a bit more flexible and stuff - patched a ton of bugs and became aware of even more bugs than i've patched - started work on command line interface for certain tools i.e chart conversions

+++ Gameplay/GameplaySnap.cs
@@ -3,14 +3,15 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using YAVSRG.Beatmap;
+using YAVSRG.Charts;
+using YAVSRG.Charts.YAVSRG;
 
 namespace YAVSRG.Gameplay
 {
     public class GameplaySnap : Snap
     {
         public int[] colors;
-        public GameplaySnap(float offset, int taps, int holds, int middles, int ends, int mines): base(offset, taps, holds, middles, ends, mines)
+        public GameplaySnap(float offset, ushort taps, ushort holds, ushort middles, ushort ends, ushort mines): base(offset, taps, holds, middles, ends, mines)
         {
             colors = new int[10];
         }
@@ -22,7 +23,7 @@ namespace YAVSRG.Gameplay
 
         public override OffsetItem Interpolate(float time)
         {
-            return new GameplaySnap(time, 0, 0, holds.value + middles.value, 0, 0);
+            return new GameplaySnap(time, 0, 0, (ushort)(holds.value + middles.value), 0, 0);
         }
     }
 }
